univerchain/requirements.txt
Django==3.2.4
web3==5.20.0
PyYAML==5.4.1
Pillow==7.0.0
isort==5.9.1
flake8==3.9.2
psycopgy2
uwsgi
djangoresframework
django-debug-toolbar
pytest-django